--- FusionTransformer/models/STN.py
@@ -1,38 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-class STN(nn.Module):
-    def __init__(self, in_channels):
-        super(STN, self).__init__()
-        # Spatial transformer localization-network
-        self.localization = nn.Sequential(
-            nn.Conv2d(in_channels, 8, kernel_size=7),
-            nn.MaxPool2d(2, stride=2),
-            nn.ReLU(True),
-            nn.Conv2d(8, 10 * 3 * 3, kernel_size=5),
-            nn.MaxPool2d(2, stride=2),
-            nn.ReLU(True),
-            nn.AdaptiveAvgPool2d(1)
-        )
-
-        # Regressor for the 3 * 2 affine matrix
-        self.fc_loc = nn.Sequential(
-            nn.Linear(10 * 3 * 3, 32),
-            nn.ReLU(True),
-            nn.Linear(32, 3 * 2)
-        )
-        # Initialize the weights/bias with identity transformation
-        self.fc_loc[2].weight.data.zero_()
-        self.fc_loc[2].bias.data.copy_(torch.tensor([1, 0, 0, 0, 1, 0], dtype=torch.float))
-    
-    def forward(self, x, output_shape):
-        B, C, H, W = x.shape
-        xs = self.localization(x)
-        xs = xs.view(-1, 10 * 3 * 3)
-        theta = self.fc_loc(xs)
-        theta = theta.view(-1, 2, 3)
-
-        grid = F.affine_grid(theta, (B, *output_shape),align_corners=False)
-        x = F.grid_sample(x, grid, align_corners=False)
-        return x
