+++ FusionTransformer/models/build.py
@@ -2,7 +2,7 @@ from FusionTransformer.models.late_fusion import LateFusionTransformer
 from FusionTransformer.models.middle_fusion import MiddleFusionTransformer
 from FusionTransformer.models.early_fusion import EarlyFusionTransformer
 from FusionTransformer.models.lidar_model import LidarSeg
-from FusionTransformer.models.image_model import ImageSeg
+from FusionTransformer.models.image_models import ImageSeg
 from FusionTransformer.models.metric import SegIoU
